POWERDK: Add power analysis support(powerdk)
BZ: 10189

Powerdk is a power driver which is the basis of the wuwatch
(wake-up watch) power analysis tool in VTune Amplifier XE that
monitors the power draw of the system. This tool is only for
internal use.

powerdk is only enabled when all of following options are set
to "y"
    CONFIG_MODULES
    CONFIG_MODULE_UNLOAD
    CONFIG_TRACEPOINTS
    CONFIG_FRAME_POINTER
    CONFIG_TIMER_STATS

Change-Id: I5423debec67ee5833334422c37f61cc64a1cbf98
Signed-off-by: Zhang Dongxing <dongxing.zhang@intel.com>
Reviewed-on: http://android.intel.com:8080/21880
Reviewed-by: Liu, Chuansheng <chuansheng.liu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>