Integrate external version of SoCWatch v1.3 driver
This commit replaces the older version of the SoCWatch driver
(v1.2) with the latest version (v1.3). SoCWatch is a tool that
enables fine-grained power optimization. SoCWatch v1.3 is a major
improvement over previous versions and includes support
for HSW, unlimited collections, 1 msec polling, and much more.
It is strongly recommended that users switch from v1.2 to v1.3
ASAP.

Issue: ADEV-457

Change-Id: I8e35b591d68bc55d8b9c2af00717368b40659926
Signed-off-by: Gautam Upadhyaya <gautam.upadhyaya@intel.com>